+++ src/Components/CategoryGrid/CategoryCard/CategoryCard.jsx
@@ -1,0 +1,11 @@
+import StyledCategoryCardContainer from "./StyledCategoryCard";
+
+const CategoryCard = ({name="the name"}) => {
+    return(
+        <StyledCategoryCardContainer>
+            <p>{name}</p>
+        </StyledCategoryCardContainer>
+    );
+}
+
+export default CategoryCard;
