+++ src/Components/CategoryGrid/StyledCategoryGrid.jsx
@@ -1,0 +1,22 @@
+import styled from "styled-components";
+
+const StyledCategoryContainer = styled.div`
+    display: flex;
+    flex-direction: column;
+    text-align: center;
+    background-color: #eaeded;
+    padding: 2rem 0;
+
+    h2{
+        padding: 1rem 0;
+    }
+    .categories-grid{
+        display: flex;
+        flex-wrap: wrap;
+        flex-direction: row;
+        justify-content: space-evenly;
+    }
+
+`;
+
+export default StyledCategoryContainer;
