+++ src/Components/Footer/Footer.jsx
@@ -1,0 +1,14 @@
+import StyledFooterContainer from "./StyledFooter";
+
+const Footer = () => {
+    return(
+        <StyledFooterContainer>
+            
+            <p className='footerText'>
+                Ecommerce created during Wizeline’s Academy React Bootcamp hola
+            </p>
+        </StyledFooterContainer>
+    );
+}
+
+export default Footer;
