+++ src/Components/ProductsGrid/ProductsGrid.jsx
@@ -1,24 +1,7 @@
 import {useEffect, useState} from 'react';
-import styled from 'styled-components';
-import Productslist from '../utils/mocks/en-us/featured-products.json';
-import ProductCard from './ProductCard';
-
-const ProductGridContainer = styled.div`
-    padding: 2rem 0;
-    display: flex;
-    flex-direction: column;
-    h2{
-        padding: 1rem 0;
-        text-align: center;
-    }
-
-    .products-grid{
-        display: flex;
-        flex-direction: row;
-        flex-wrap: wrap;
-        justify-content: space-evenly;
-    }
-`;
+import Productslist from '../../utils/mocks/en-us/featured-products.json';
+import ProductCard from './ProductCard/ProductCard';
+import ProductGridContainer from './ProductsGridStyled';
 
 const ProductGrid = () => {
