+++ src/Pages/Shop.jsx
@@ -1,5 +1,5 @@
-import Header from '../Components/Header';
-import Footer from '../Components/Footer';
+import Header from '../Components/Header/Header';
+import Footer from '../Components/Footer/Footer';
 import Homepage from './Homepage';
 import styled from 'styled-components';
